requirements.txt
alembic==1.7.7
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
dnspython==2.2.1
email-validator==1.1.3
factory-boy==3.2.1
Faker==13.3.3
fastapi==0.73.0
freezegun==1.2.1
gprof2dot==2021.2.21
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.2.0
httpx==0.22.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
Mako==1.2.0
mangum==0.14.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.6.7
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
pprintpp==0.4.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pydocstyle==6.0.0
Pygments==2.11.2
pyinstrument==4.1.1
pyparsing==3.0.7
pytest==7.1.1
pytest-asyncio==0.18.3
pytest-clarity==1.0.1
pytest-factoryboy==2.1.0
pytest-mock==3.7.0
pytest-profiling==1.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
regex==2021.3.17
requests==2.27.1
rfc3986==1.5.0
rich==12.0.1
s3transfer==0.5.2
six==1.16.0
snakeviz==2.1.1
sniffio==1.2.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a21
sqlmodel==0.0.6
sqlparse==0.4.2
sqltap==0.3.11
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
ujson==4.3.0
urllib3==1.26.9
watchgod==0.8.1
websockets==10.2
Werkzeug==2.0.3